\name{PCIT-package}
\alias{PCIT-package}
\docType{package}
\title{
Partial Correlation coefficient with Information Theory (PCIT)
}
\description{
  This package provides the necessary functions for performing the
  Partial Correlation coefficient with Information Theory (PCIT) algorithm
  developed by Reverter and Chan (2008). The PCIT algorithm identifies
  significant gene to gene associations to define edges in a weighted network.
}
\details{
\tabular{ll}{
  Package: \tab PCIT\cr
  Type: \tab Package\cr
  Version: \tab 1.01-1\cr
  Date: \tab 2009-08-25\cr
  License: \tab GPL-3\cr
  LazyLoad: \tab yes\cr
}
}
\author{
Nathan S. Watson-Haigh

Maintainer: Nathan S. Watson-Haigh <nathan.watson-haigh@csiro.au>
}
\references{
  A. Reverter and E.K.F. Chan. (2008) Combining partial correlation and an information theory approach to the reversed engineering of gene co-expression networks. Bioinformatics. \strong{24}(21), 2491-2497.
}
\keyword{ package }
