\name{PCIT-package}
\alias{PCIT-package}
\alias{PCIT}
\docType{package}
\title{
	Partial Correlation coefficient with Information Theory (PCIT)
}
\description{
	This package provides the necessary functions for performing the
	Partial Correlation coefficient with Information Theory (PCIT) algorithm
	developed by Reverter and Chan (2008). The PCIT algorithm identifies
	meaningful correlations to define edges in a weighted network.
	
	The algorithm can be applied to any correlation-based network including but
	not limited to gene co-expression networks.
}
\details{
\tabular{ll}{
	Package: \tab PCIT\cr
	Type: \tab Package\cr
	Version: \tab 1.04-3\cr
	Date: \tab 2011-05-11\cr
	License: \tab GPL-3\cr
	LazyLoad: \tab yes\cr
}
}
\author{
	Nathan S. Watson-Haigh
	Maintainer: Nathan S. Watson-Haigh <nathan.watson-haigh@awri.com.au>
}
\references{
	A. Reverter and E.K.F. Chan. (2008) Combining partial correlation and an information theory approach to the reversed engineering of gene co-expression networks. Bioinformatics. \strong{24}(21), 2491-2497.
}
\keyword{ package }
